+++ smart_fruit/feature_types/__init__.py
@@ -1,0 +1,5 @@
+from smart_fruit.feature_types.feature_type_base import FeatureType
+from smart_fruit.feature_types.simple_types import Number, Integer, Complex, Label, Tag
+from smart_fruit.feature_types.compound_types import Vector
+
+__all__ = ["FeatureType", "Number", "Integer", "Complex", "Label", "Vector", "Tag"]