+++ smart_fruit/model.py
@@ -1,11 +1,11 @@
-from csv import DictReader
-
 from pandas import DataFrame, concat
+from pandas.core.apply import frame_apply
 
 from sklearn import linear_model
 
-from smart_fruit.feature_types import FeatureClassMeta
+from smart_fruit.feature_class import FeatureClassMeta
 from smart_fruit.model_selection import train_test_split
+from smart_fruit.utils import csv_open
 
 __all__ = ["Model"]
 
@@ -44,8 +44,7 @@ class Model(metaclass=ModelMeta):
 
     @classmethod
     def input_features_from_csv(cls, csv_path):
-        with open(csv_path) as csv_file:
-            yield from cls.input_features_from_json(DictReader(csv_file))
+        yield from cls.input_features_from_json(csv_open(csv_path, cls.Input._fields))
 
     @classmethod
     def features_from_list(cls, lists):
@@ -59,8 +58,7 @@ class Model(metaclass=ModelMeta):
 
     @classmethod
     def features_from_csv(cls, csv_path):
-        with open(csv_path) as csv_file:
-            yield from cls.features_from_json(DictReader(csv_file))
+        yield from cls.features_from_json(csv_open(csv_path, cls.Input._fields + cls.Output._fields))
 
     @staticmethod
     def _to_raw_features(dataframe, feature_class):
@@ -78,12 +76,13 @@ class Model(metaclass=ModelMeta):
         return input_dataframe, output_dataframe
 
     @classmethod
-    def train(cls, features, train_test_split_ratio=None, test_sample_count=None):
+    def train(cls, features, train_test_split_ratio=None, test_sample_count=None, random_state=None):
         if train_test_split_ratio is not None or test_sample_count is not None:
             train_features, test_features = train_test_split(
                 features,
                 train_test_split_ratio=train_test_split_ratio,
-                test_sample_count=test_sample_count
+                test_sample_count=test_sample_count,
+                random_state=random_state
             )
 
             model = cls.train(train_features)
@@ -103,18 +102,31 @@ class Model(metaclass=ModelMeta):
     def _chunk_dataframe(dataframe, feature_types):
         start = 0
         for feature_type in feature_types:
-            yield dataframe.loc[:, start:start + feature_type.feature_count - 1], feature_type
+            chunk = dataframe.loc[:, start:start + feature_type.feature_count - 1]
+            chunk.columns = range(len(chunk.columns))
+
+            yield chunk, feature_type
             start += feature_type.feature_count
 
-    def predict(self, input_features):
-        raw_features = self._to_raw_features(DataFrame(input_features), self.Input)
+    def predict(self, input_features, yield_inputs=False):
+        input_features_dataframe = DataFrame(input_features)
+
+        raw_features = self._to_raw_features(input_features_dataframe, self.Input)
 
         raw_prediction_dataframe = DataFrame(self.model.predict(raw_features))
 
+        # Using frame_apply instead of chunk.apply as chunk.apply doesn't behave as expected for 0 columns
         prediction_dataframe = concat([
-            chunk.apply(feature_type.from_series, axis=1)
+            frame_apply(chunk, feature_type.from_series, axis=1).apply_standard()
             for chunk, feature_type in self._chunk_dataframe(raw_prediction_dataframe, self.Output)
         ], axis=1)
 
-        for _, output_series in prediction_dataframe.iterrows():
-            yield self.Output(*output_series)
+        if yield_inputs:
+            for (_, input_series), (_, output_series) in zip(
+                input_features_dataframe.iterrows(),
+                prediction_dataframe.iterrows()
+            ):
+                yield self.Input(*input_series), self.Output(*output_series)
+        else:
+            for _, output_series in prediction_dataframe.iterrows():
+                yield self.Output(*output_series)